services/auth_service/requirements.txt
fastapi
uvicorn
beanie
dotenv
redis
pymongo>=4.9
pyjwt
